requirements.txt
streamlit
tika
PyMuPDF
faiss-cpu
requests
langchain
huggingface-hub
transformers
sentence-transformers